requirements.txt
Django
djangorestframework
django-cors-headers
celery
redis
pandas
openpyxl
django-filter
psycopg2
psycopg2[binary]
python-dotenv